+++ db.py
@@ -1,76 +1,75 @@
-import sqlite3
+import mariadb as db
+import sys
 
-connect_mes = "Successfully Connected to SQLite database wxdata.db"
-close_mes = "The SQLite connection is closed"
-insert_mes = "Record inserted successfully into"
+def create_db():
+    db_init = """ CREATE DATABASE IF NOT EXISTS weather;"""
+    weather_table_init = """ USE weather;
+                CREATE TABLE IF NOT EXISTS sensors (
+                ID BIGINT NOT NULL PRIMARY KEY AUTO_INCREMENT,
+                stationid VARCHAR(10),
+                created TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+                ambient_temperature DECIMAL(6,2) NOT NULL,
+                wind_direction DECIMAL(6,2),
+                wind_speed DECIMAL(6,2),
+                wind_gust_speed DECIMAL(6,2),
+                humidity DECIMAL(6,2) NOT NULL,
+                rainfall DECIMAL(6,2),
+                air_pressure DECIMAL(6,2) NOT NULL,
+                PM25 DECIMAL(6,2),
+                PM10 DECIMAL(6,2)
+                );"""
+    packet_table_init = """ USE weather;
+                            CREATE TABLE IF NOT EXISTS packets(
+                            ID BIGINT NOT NULL PRIMARY KEY AUTO_INCREMENT,
+                            created TIMESTAMP DEFAULT CURRENT_TIMESTAMP, 
+                            packet VARCHAR(100) NOT NULL,
+                            transmitted BOOL NOT NULL
+                            );"""
 
-def make_table():
-    sqlite_insert_query1 = """ CREATE TABLE IF NOT EXISTS weather(
-                                ID INTEGER PRIMARY KEY,
-                                DateTime TEXT NOT NULL,
-                                StationID TEXT,
-                                TemperatureF NUMERIC NOT NULL,
-                                Pressure NUMERIC NOT NULL,
-                                Humidity NUMERIC NOT NULL,
-                                PM25 NUMERIC,
-                                PM10 NUMERIC,
-                                AQI
-                            );"""
-    sqlite_insert_query2 = """ CREATE TABLE IF NOT EXISTS packets(
-                                ID INTEGER PRIMARY KEY,
-                                DateTime TEXT NOT NULL,
-                                packet TEXT NOT NULL,
-                                Sent INTEGER NOT NULL
-                            );"""
     try:
-        sqliteConnection = sqlite3.connect('wxdata.db')
-        cursor = sqliteConnection.cursor()
-        print(connect_mes)
-        cursor.execute(sqlite_insert_query1)
-        print(f"weather table successfully created")                                                                        
-        sqliteConnection.commit()
-        cursor.execute(sqlite_insert_query2)
-        sqliteConnection.commit()
-        print(f"{cursor.rowcount} packets table successfully created")                                                                        
-    except sqlite3.Error as error:
-            print(f"CRITICAL: Failed to create weather table: {error}")
-    finally:
-            if (sqliteConnection):
-                sqliteConnection.close()
-               # print("The SQLite connection is closed")
+        conn = db.connect(
+            user="wxstation",
+            password="password1",
+            host="127.0.0.1",
+            port=3306
+        )
+        cur = conn.cursor()
+        cur.execute(db_init); conn.commit()
+        cur.execute(weather_table_init); conn.commit()
+        cur.execute(packet_table_init); conn.commit()
+        conn.close()
+    except db.Error as e:
+        print(f"Error connecting to MariaDB Server: {e}")
+        sys.exit(1)
 
-def read_save_enviro(data):
-  try:
-        sqliteConnection = sqlite3.connect('wxdata.db')
-        cursor = sqliteConnection.cursor()
-        weather_insert = """INSERT INTO weather(DateTime, StationID, TemperatureF, Pressure, Humidity, pm25, pm10) 
-        VALUES(?, ?, ?, ?, ?, ?, ?);"""
-        data_tuple = ("datetime('now')", data['callsign'], data['temperature'], data['pressure'], data['humidity'], data['pm25'], data['pm10'])
-        cursor.execute(weather_insert, data_tuple)
-        sqliteConnection.commit()
-        print(f"{cursor.rowcount} {insert_mes} weather table,", end=" ")
-        cursor.close()
-  except sqlite3.Error as error:
-        print(f"CRITICAL: Failed to insert data into sqlite weather table: {error}")
-  finally:
-        if (sqliteConnection):
-            sqliteConnection.close()
+def db_connect():
+    try:
+        conn = db.connect(
+            user="wxstation",
+            password="password1",
+            host="127.0.0.1",
+            port=3306,
+            database="weather"
+        )
+        return conn.cursor()
+    except db.Error as e:
+        print(f"Error connecting to MariaDB Server: {e}")
+        sys.exit(1)
+
+def read_save_sensors(data):
+    weather_insert = """INSERT INTO weather(StationID, TemperatureF, Pressure, Humidity, pm25, pm10) 
+    VALUES(?, ?, ?, ?, ?, ?);"""
+    data_tuple = (data['callsign'], data['temperature'], data['pressure'], data['humidity'], data['pm25'], data['pm10'])
+    conn = db_connect()
+    cur = conn.cursor()
+    cur.execute(weather_insert, data_tuple)
+    conn.commit(); conn.close()
 
 def read_save_packet(data):
-    try:
-        sqliteConnection = sqlite3.connect('wxdata.db')
-        cursor = sqliteConnection.cursor()
-        packet_insert = """INSERT INTO packets(DateTime, packet, Sent) 
-        VALUES(?, ?, ?);"""
-        data_tuple = ("datetime('now')", data['packet'], data['sent'])
-        cursor.execute(packet_insert, data_tuple)
-        sqliteConnection.commit()
-        print(f"{cursor.rowcount} {insert_mes} packet table")
-        cursor.close()
-
-    except sqlite3.Error as error:
-          print(f"CRITICAL: Failed to insert data into sqlite packets table: {error}")
-    finally:
-        if (sqliteConnection):
-          sqliteConnection.close()
-          #print(close_mes)
+    packet_insert = """INSERT INTO packets(packet, Sent) 
+    VALUES(?, ?);"""
+    data_tuple = (data['packet'], data['sent'])
+    conn = db_connect()
+    cur = conn.cursor()
+    cur.execute(packet_insert, data_tuple)
+    conn.commit(); conn.close()
